fe lsp sub command
