+++ crates/fe/Cargo.toml
@@ -19,7 +19,9 @@ fs_extra = "1.2.0"
 walkdir = "2"
 indexmap = "1.6.2"
 include_dir = "0.7.2"
+tokio = { version = "1.17.0", features = ["full"] }
 
+fe-lsp = {path = "../fe-lsp", version = "^0.1.0"}
 fe-common = {path = "../common", version = "^0.18.0-alpha"}
 fe-driver = {path = "../driver", version = "^0.18.0-alpha"}
 fe-parser = {path = "../parser", version = "^0.18.0-alpha"}
